Refactor Company Day analytics handling: consolidate context normalization and event emission in CompanyDayAnalyticService
